refactor: reimplement all directive utils and classes
- Add `Directive` abstract class for managing attribute-based directives.
- Introduce `LitDirective` subclass for rendering with `lit-html`.
- Create `lazyDirective` function for lazy registration of directives.
- Refactor queue rendering logic into a centralized batching system.
- Remove deprecated utility decorators and streamline directive exports.
